requirements.txt
flask>=2.3.0
flask-sqlalchemy>=3.0.0
flask-cors>=4.0.0
flask-limiter>=3.5.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
werkzeug>=2.3.0
pyjwt>=2.8.0
gunicorn>=21.0.0
cryptography>=42.0.0
